Fix up leibniz and refactor goal handling
It compiles and runs! Doesn't get the right answer yet though. Also most
of the rules are still missing.

It doesn't use an explicit goal anymore. Instead when the first rule set
fails to match, the control system switches to the next rule set and so
on. A more sophisticated system might be nice as an extension, but this is
pretty simple.